requirements.txt
openai>=1.0.0
PyYAML>=6.0.1
tqdm>=4.66.0 